backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
SQLAlchemy==2.0.36
psycopg[binary]==3.2.12
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-jose==3.3.0
email-validator==2.1.1
pydantic==2.10.4
